Break apart Menu's test suites
